config file for both networks added, as well as commands for running them added
