+++ package.json
@@ -7,6 +7,7 @@
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
     "@walletconnect/web3-provider": "^1.4.1",
+    "dotenv-cli": "^4.0.0",
     "ethers": "^5.3.1",
     "node-sass": "^4.9.4",
     "react": "^17.0.2",
@@ -17,6 +18,8 @@
   },
   "scripts": {
     "start": "react-scripts start",
+    "start:testnet": "dotenv -e .env react-scripts start",
+    "start:mainnet": "dotenv -e .env react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
