--- src/configs/mainnet.js
@@ -1,12 +1,0 @@
-const config = {
-    "network": {
-        "symbol": "ETH",
-        "name": "Ethereum Mainnet",
-        "network": "mainnet",
-        "chain_id": 1,
-        "network_id": 1,
-       // "rpc_url": `https://mainnet.infura.io/v3/${process.env.REACT_APP_INFURA_ID}`,
-    }
-}
-
-module.exports = config;
