--- src/configs/testnet.js
@@ -1,11 +1,0 @@
-const config = {
-    "network": {
-        "symbol": "ETH",
-        "name": "Ethereum Ropsten",
-        "network": "ropsten",
-        "chain_id": 3,
-        "network_id": 3,
-       // "rpc_url": `https://ropsten.infura.io/v3/${process.env.REACT_APP_INFURA_ID}`,
-    },
-}
-module.exports = config;
